requirements.txt
Django==5.2.11
tensorflow-cpu==2.16.2
keras==3.12.1
numpy==1.26.4
pillow==12.1.1
gunicorn==23.0.0
requests==2.32.5
h5py==3.15.1
asgiref==3.11.1
sqlparse==0.5.5
typing_extensions==4.15.0
